requirements-dev.txt
pytest
pytest-django
pytest-sugar
pytest-cov
coverage
